% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw_data_introduction.R
\docType{data}
\name{results}
\alias{results}
\alias{demographic1}
\alias{demographic2}
\title{Raw data used in introduction data recipe}
\format{Three objects of type \code{data.frame} named results, demographic1 and demographic2.}
\usage{
data(introduction_raw)
}
\description{
For further details see \code{learnr::run_tutorial("en-introduction", package = "ddj")}
}
\examples{
data(introduction_raw)
}
\keyword{datasets}
